Change server responses to be written using the method's unmodified name
Summary:
The reason for this diff is to avoid needing to pointlessly pass around 2 separate function name arguments in `fbthrift::help::serialize_result_envelope` in D57594686, one for the `write_message_begin` (`"{{function:name}}"`) and one for the `write_struct_begin` (originally `"{{function:upcamel}}"`).

**This makes no difference.**

Today the only protocols which look at the `name` string passed to `write_struct_begin` are:

- Capella's `ConstValueWriter`
https://www.internalfb.com/code/fbsource/[7cf0ad11bf6d70926966004d360e163c7de5c50a]/fbcode/datainfra/privacy/ppf/purpose_policy/rust/capella_const_value_protocol/src/writer.rs?lines=177-188

- `DeserializeToValueProtocol`
https://www.internalfb.com/code/fbsource/[7cf0ad11bf6d70926966004d360e163c7de5c50a]/fbcode/host_management/service_configs/libcfgen/rust/src/lean_thrift/deser_to_value.rs?lines=79-86

- MetalOS's "starlark util"
https://www.internalfb.com/code/fbsource/[7cf0ad11bf6d70926966004d360e163c7de5c50a]/fbcode/metalos/lib/starlark_util/src/value.rs?lines=203-208

- SysInspector's `ValueWriter`
https://www.internalfb.com/code/fbsource/[7cf0ad11bf6d70926966004d360e163c7de5c50a]/fbcode/sysinspector/aggregator/engine/thrift.rs?lines=118-121

As seen in {D57630441} and {D57630442}, there is no evidence of Thrift-generated Exn and StreamExn types being serialized to any of the above protocols.

The only protocol that an Exn type is ever serialized to is `SimpleJsonProtocol`, which ignores this name argument.

https://www.internalfb.com/code/fbsource/[7cf0ad11bf6d70926966004d360e163c7de5c50a]/fbcode/service_capacity/srm/status_service/src/handlers/status_service.rs?lines=192

https://www.internalfb.com/code/fbsource/[7cf0ad11bf6d70926966004d360e163c7de5c50a]/fbcode/thrift/lib/rust/src/simplejson_protocol.rs?lines=176-181

The compact and binary protocol also obviously ignore the struct name.

Reviewed By: shayne-fletcher

Differential Revision: D57594687

fbshipit-source-id: e16ef7a5db63307d5d3e17066d1a5c5729ee6e60