ServiceInterceptorStorage
Summary:
This class represents the underlying storage of the `RequestState` template param in the public ServiceInterceptors API (see D55162055). 64 bytes is a fairly low number that we can tune later if necessary.

The memory allocated per interceptor instance must be static so that we can efficiently compute offsets to locate where the storage is in memory, given the index (order) of the `ServiceInterceptor` implementation.

Reviewed By: iahs

Differential Revision: D55044945

fbshipit-source-id: 3566ecec086106503d49f2680830cf2972d01c9c