Remove unused method (Get<Foo>NonCompat)
Summary:
The idea behind this `Get<Foo>NonCompat()` method was to replace the legacy `Get<Foo>()` method one day.

There are a number of problems with the legacy `Get<Foo>()` - for example, for `optional` unset lists/maps/set fields - the method would return an empty slice/map, rather than `nil`. This is unexpected, misleading and error-prone - an unset `optional` field should remain `nil` - otherwise it's impossible to tell if it's set or unset. There are a few other non-obvious behaviors like that.

...In any case: we've had this method for a while, but not enough time or resources to invest into a proper migration. Plus, I think the approach to getters should be modified a bit (e.g. there needs to be a `Get<Foo>OrDefault()` method to allow callers to specify a default value in case the field is unset). TLDR - the entire approach to getters needs to be thought through a bit more.

In the meantime - the extra code produced by this method is contributing to codegen volume and slower builds. So, for the foreseeable future - let's remove it.

Reviewed By: leoleovich

Differential Revision: D57501011

fbshipit-source-id: 37d13f037f3ac8a1c2487e4609900e3f39f029fd