Remove the deprecated visitor from Thrift
Summary: Now that all but one use of the deprecated Thrift visitor API have been migrated, remove it from Thrift by relocating to the detail namespace in the only usage site. There is no change in behavior since the implementation of the visitor remains the same. We can separately remove most of the logic (which will require more careful review) since this usage site doesn't need all of the functionality.

Reviewed By: Mizuchi

Differential Revision: D57746085

fbshipit-source-id: c90e2ace1370fe234112ea72509cd043f24ffd38