Optimize register_method to avoid double-boxing function
Summary:
In D57694787 I was not able to make `register_methods` delegate to `register_method` because that regressed performance — the argument that comes into `register_method` gets boxed as a `std::function`, and then boxed again with a different signature to be stored in `methods`.

This diff removes the boxing of `register_method`'s argument. Every function passed to any of the `register_…` functions will be boxed exactly once. I confirmed that this is now performance-neutral with `register_methods` delegating to `register_method`.

Reviewed By: shayne-fletcher

Differential Revision: D57743749

fbshipit-source-id: f60245bbe59837edbaea5d5609bf32aee842e134