Turn on caching of mstch nodes
Summary:
The Java Thrift implementation relies on quite a bit of mutation in mstch methods.

https://www.internalfb.com/code/fbsource/[04316cb722ee5a091574e65a2074640072bf1a09]/fbcode/thrift/compiler/generate/t_mstch_java_generator.cc?lines=718-722%2C735%2C738%2C750-757

I have left those as `register_methods` in this diff. This will be renamed to `register_volatile_methods` in D57785398 to indicate that every place `{{function:nestedDepth++}}` is mentioned in the templates, the call actually needs to be performed, it cannot reuse a cached value.

Reviewed By: iahs

Differential Revision: D57785371

fbshipit-source-id: c17c2cce881232b332b567799844a64e7733fed4