Fix a nestedDepth leak in adapter toThrift conversion
Summary:
A Java build failure got bisected to D57765393. Fortunately that diff is easy to bisect further. I ascertained that specifically the change in `mstch_struct` causes the failure, and even within that, specifically the `{{#struct:fields}}` method. Converting the single method `{{#struct:fields}}` back from `register_cached_method` to `register_volatile_method` unbreaks the target.

At first this seemed strange because `{{#struct:fields}}` is a conceptually simple getter, it returns an array of the struct's fields. There should be no mutation involved; every call should be returning the same fields every time, so caching the call should not be an issue.

https://www.internalfb.com/code/fbsource/[6bf38b63c308e4df3cd655fa97bc199367c3ff97]/fbcode/thrift/compiler/generate/mstch_objects.cc?lines=386-388

But while `struct_->get_members()` is pure and returns the same value on every call, `make_mstch_fields` is not.

https://www.internalfb.com/code/fbsource/[6bf38b63c308e4df3cd655fa97bc199367c3ff97]/fbcode/thrift/compiler/generate/mstch_objects.h?lines=980-990

It runs the language's `field_factory` to allocate a **new** `mstch_field` for each field in the struct. If Java's `mstch_field` were immutable like every other language's, there still would be no issue with caching. But it's not; the difference between caching `make_mstch_fields` and not caching it is whether mutations made on the fields returned by an earlier call are visible in the fields returned by a subsequent call, or whether each call produces a fresh collection of `mstch_field`.

Java's `mstch_field` holds a mutable `nestedDepth` that can be incremented and decremented from a mstch template. It is used for generating a sequence of distinct variable names such as `_value{{field:nestedDepth}}` because Java does not support shadowing variable names in nested scopes (*"error: variable x is already defined in method"*).

https://www.internalfb.com/code/fbsource/[b3581dec92742580dde627f7c9196f7fcbdbc915]/fbcode/thrift/compiler/generate/t_mstch_java_generator.cc?lines=772%2C823-828%2C834%2C837%2C929%2C937-944

So if anywhere in the Java templates performs a mismatched number of calls to `{{field:nestedDepth++}}` and `{{field:nestedDepth--}}`, it will leave a `mstch_java_field` with a `nestedDepth` value different from its default value of 0. Then subsequent calls of `{{#struct:fields}}`, which used to produce only freshly constructed `mstch_java_field` instances having nestedDepth=0, after caching is applied they would return previously created and mutated `mstch_java_field` having nonzero nestedDepth.

This possibility is reinforced by the difference in generated code before and after the caching diff. It starts off with:

```lang=diff
-                Set<Integer> _value1;
+                Set<Integer> _value2;
             {
-            TSet _set1 = oprot.readSetBegin();
-            _value1 = new HashSet<Integer>(Math.max(0, _set1.size));
-            for (int _i1 = 0; (_set1.size < 0) ? oprot.peekSet() : (_i1 < _set1.size); _i1++) {
+            TSet _set2 = oprot.readSetBegin();
+            _value2 = new HashSet<Integer>(Math.max(0, _set2.size));
+            for (int _i2 = 0; (_set2.size < 0) ? oprot.peekSet() : (_i2 < _set2.size); _i2++) {
```

i.e. some variable names off by 1. But further in the diff, the variable names get increasingly far apart.

```lang=diff
-                int _value5 = oprot.readI32();
+                int _value8 = oprot.readI32();

-                _value4.add(__adaptedInt_Adapter.fromThrift(_value5));
+                _value7.add(__adaptedInt_Adapter.fromThrift(_value8));
```

The missing `{{field:nestedDepth--}}` was easy to find and this appears to be the only one.

```
$ rg 'field:nestedDepth\+\+' thrift/compiler/generate/templates/java | wc -l
50

$ rg 'field:nestedDepth--' thrift/compiler/generate/templates/java | wc -l
49
```

Reviewed By: zertosh

Differential Revision: D57809171

fbshipit-source-id: 0968f8dd5f6128bd622a65168782fa50cdb9f752