Don't output cython files when compiling in auto migrate mode
Summary:
The whole py3_types.pyx/types.pyx split has been removed, now just if we're in auto_migrate mode we generate types.py and re-export thrift python (and the same for clients/services/etc), so no longer generating these files in the mustache compiler and the equivalent buck macro.

The only special files are `types.pxd`, which exists for re-exporting c bindings, and `builders.py`, which exists for builders. I don't think reflection files have to be generated but I'll need to check on that, they're excluded from generation for now but I'll check back on that.

Reviewed By: ahilger

Differential Revision: D57939633

fbshipit-source-id: e6d32e7a21db37784cc5ca91200ae24d2cc915af