Delete support for delimiter change token {{= ... =}}
Summary:
The deleted test showcases how this feature worked.

https://www.internalfb.com/code/fbsource/[a7dff46bded8e0f3f4d85e0c8384c77070f02609]/fbcode/thrift/compiler/detail/mustache/test/delimiters_test.cpp?lines=32-36

The syntax `{{=`...`=}}` means split `...` by space and use its first word as the new open delimiter for the rest of this template instead of `{{`, and the second word as the close delimiter instead of `}}`.

This is not a feature we use, and it contributes 8% performance cost in having to do bookkeeping for the chosen delimiters at the template, section, and token level.

Reviewed By: iahs

Differential Revision: D57946620

fbshipit-source-id: 9e6673999778cd5939c41ad37f3b61bd74086189