Execute ServiceInterceptorBase::onRequest
Summary:
**Meaningful changes under `thrift/lib` and `thrift/compiler/generate`**.

This diff implements `ServiceInterceptor::onRequest` functionality. This should happen after deserialization so the invocation needs to be in generated code (since deserialization happens in generated code).

The change in control flow is quite simple — before invoking `async_tm_*` or `async_eb_*` in the generated `executeRequest_*` function, we asynchronously run the `ServiceInterceptor::onRequest` implementations in order.

To make the generated code simpler, we create a function-scoped struct to store request arguments so that we can use the implicitly-defined move constructor rather than moving each argument by hand.

If there are any exceptions thrown in `onRequest`, we capture and collect them. This means that `onRequest` for *all* instances are called even if there is an exception. However, the execution short-circuits and sends an error back to the client if there is an exception without executing the service handler code.

Reviewed By: iahs

Differential Revision: D55162054

fbshipit-source-id: bddf0ace65f313cfd3ee92cb43ebb497785a69c8