Add useCount() to the default control block implementation in IntrusiveSharedPtr
Summary: The use count in the next diff to check if someone has called `unsafeRelease()` on the shared pointer (i.e. use count is 0). If so, then the user has assumed ownership of the `HandlerCallback` object and we cannot execute `ServiceInterceptorBase::onResponse` (because it extends the request lifetime beyond the user's own code).

Reviewed By: maalbash

Differential Revision: D57595745

fbshipit-source-id: 8e9c634f5e85b85a541d5f752630a6c3016f86f4