Allow AsyncProcessorFactory to select the ResourcePool used for a request
Summary:
This change enables AsyncProcessorFactory implementations to (optionally) direct requests handled by the processor returned by `::getProcessor()` to the ResourcePool of their choice.

Meta:

It also updates the connection handlers for the "extra interfaces" (e.g. Monitoring, Debug, etc) to use the server's ResourcePoolSet.

Why this diff?

This diff is entirely in service of making it easier to add MetricCollector-based instrumentation to Resource Pools components. Importantly, it enables the invariant that there is only ever be one ResourcePoolSet per server. Conveniently, the server's ResourcePoolSet is already easily accessible for instrumentation purposes.

Backstory on Resource Pools and the "extra interfaces" over special connections:

Currently, the Monitoring, Profiling, and Debug extra interfaces execute their requests on separate ResourcePools/Executors, but only if the methods are called on a special Monitoring, Profiling, or Debug connection, respectively. If these interface methods are called on a standard connection, they will be executed on the default resource pools. This behavior is (probably) not intended. Prior to this diff, the way this ResourcePool "selection" is implemented today is a total hack that abuses the ThriftServer ServerConfigs API. This hack was needed, because there was no way to direct certain interface methods to different ResourcePools prior to this diff.

What changes when this diff lands?

Behaviorally, nothing. However, there is now a single ResourcePoolSet per ThriftServer instance that is owned by the ThriftServer. Each extra interface connection handler simply adds its own ResourcePool to the server's Set and directs requests to it.

What is also included:
* Removed the ability for Rocket ConnectionHandlers to effictively override ThriftServer behavior
* Killed ServerConfigsWrapper

What isn't included:
* `ServerConfigs::resourcePoolSet` methods still exist - to prevent future abuse, these should be removed and ThriftServer impl made non-virtual
* Internal default extra interface impls should consider implementing `::selectResourcePool()` to ensure methods get routed correctly regardless of connection type
* The extra interface executors still live in singletons and, therefore, are shared across all ThriftServer instances in the process - probably fine, but more isolation may be preferred

Reviewed By: praihan

Differential Revision: D57887250

fbshipit-source-id: c8efcc6f964a45cc29443da03f45e8380fc3b5ab