remove Flush call from header transport except when it is a Buffered Transport
Summary:
remove Flush call from header transport

There is no real Flush logic in socket
```
// for interface compatibility with Thrift1.
func (s *socket) Flush() error {
	return nil
}
```

We want to remove Flush from the Socket interface, so we can move closer to net.Conn

Unfortunately we have a BufferedTransport which is sometimes used with the HeaderProtocol, so we can't completely get rid of it. We added some tests to make sure this is not missed again in future.
https://www.internalfb.com/code/search?hide_uninteresting=false&leading_context=3&trailing_context=3&lang_filter=go&repo=fbcode&q=repo%3Afbcode%20-filepath%3A(third-party%7Cgithub_repos)%20NewBufferedTransport

Reviewed By: echistyakov

Differential Revision: D57972161

fbshipit-source-id: 7aeea080683889b9e42ff00313cd3cd4d624dc01