Serialize map keys as named enum constants
Summary:
It has been reported in [this post](https://fb.workplace.com/groups/configerator.users/permalink/3844753095756852/) that keys in Thrift maps are being serialized as integers (and not named enum values such as `EnumType.VALUE`) even when using `setNamedEnumValues(true)`. The cause for that was that the generated ttypes.py file has a separate logic for generating the part of `readFromJson()` method that deals with getting the key for setting some value in a map.

The fix is simple. I haven't looked for other places that potentially serialize enum values, but haven't found any.

Differential Revision: D58056801

fbshipit-source-id: 564d63e15f3f9f4c13710a370bc6f1033fe77d58