Deshim FunctionScheduler in thrift
Summary:
The following rules were deshimmed:
```
//folly/experimental:function_scheduler -> //folly/executors:function_scheduler
```

The following headers were deshimmed:
```
folly/experimental/FunctionScheduler.h -> folly/executors/FunctionScheduler.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:ad_defsx.thrift

Reviewed By: yfeldblum

Differential Revision: D58155075

fbshipit-source-id: ab249347e5d464ab41b5acf410dbc580945b9693