Allow disabling ResourcePool join executor on ::stop() behavior
Summary:
Users are reporting log spam resulting from D57887250 due to the use of custom, non-owning executor wrappers in the auxiliary ResourcePools.

This diff adds the ability to tell a ResourcePool that it should not attempt to join the executor on exit. This is useful for cases where multiple ResourcePools internally use the same executor.

Should the ResourcePool be configured to join the executor on stop, the original WARNING message has also been updated to include the name of the offending ResourcePool.

Differential Revision: D58291989

fbshipit-source-id: 0d04f97f2b7c7ceaa6741354acce6e2b3f8b0956