Restore writeMapValueBegin/End hooks in op::encode path
Summary:
These hooks were added to protocol_methods in D30942212 to support the "deterministic hashing" feature. They were not carried over to op::encode, which means types using this serialization path are incorrectly shuffling keys and values inside map fields before hashing. While the result is technically still a valid hash (in the same sense that a constant function is a valid hash), it means transposing values within a map or inverting it would not change the hash which makes the hash useless for verifying integrity.
This diff adds the missing hooks, which will change the value of op::hash for affected structs (including protocol.Value containing a map field).

Reviewed By: thedavekwon

Differential Revision: D58147573

fbshipit-source-id: 4d4f7b9c20fb476c01cdef6fc3f414b8aa1f7e1c