Respect customized module names in generated code
Summary:
The top-level module names in multi-file Thrift libraries are supposed to be controllable using `namespace rust cratename.modulename`. This works and produces modules with the requested name — that happens here:

https://www.internalfb.com/code/fbsource/[3131004005dc7e19f6b66cf492356c36ee4cd45d]/tools/build_defs/fbcode_macros/build_defs/lib/thrift/rust/crate_root.py?lines=65-69

And if you are using a multi-file Thrift-generated Rust library from **handwritten** Rust code, everything is good.

But **Thrift-generated** code that depends on a multi-file Thrift-generated Rust library was using incorrect module names. It would not respect the modulename written in the dependency's `namespace rust`, and would instead continue to use a module name derived from the dependency's Thrift filename.

This diff fixes the Rust Thrift implementation to correctly refer to modules that have a module name customized by a `namespace rust` statement.

Reviewed By: zertosh

Differential Revision: D58398525

fbshipit-source-id: f232802612ab9985cb441f3e9135781c8b1f521a