Use folly::exception_wrapper to save exceptions in ScopedServerThread
Summary:
`ScopedServerThread` has an allowlist of exceptions that it understands and can handle. However, with the help of `std::exception_ptr` and `folly::exception_wrapper`, we can automatically handle all exception types!

There is a subtle change in behavior here — exceptions previously unrecognized would become `std::exception` but now will be preserved. This class is primarily used in tests so should not be an issue in practice (if this case even pops up in the first place).

I also discovered that Thrift server's runtime does not call `TServerEventHandler::handleServerError` anywhere. The Haskell Thrift server seems to re-use the type so it can't be immediately removed but should be feasible with a small (but separate diff).

Reviewed By: iahs

Differential Revision: D58267040

fbshipit-source-id: 5bec272c1df921e2911c10a42f8f47c8b7e12e21