Add ServiceInterceptor::getName
Summary:
This adds the requirement for `ServiceInterceptor` implementations to have names. Now they form a qualified name of the form `"{module_name}.{interceptor_name}"`.

This allows us to detect duplicates and (in a later diff) include the interceptor name in error messages from exceptions thrown from a `ServiceInterceptor` implementations.

I added an extra check in `ScopedServerThread` to detect `std::logic_error` for my test case. The whole `SavedException` machinery is super jank and can be replaced with `folly::exception_wrapper`. I plan to do that in a future diff.

Reviewed By: iahs

Differential Revision: D58262630

fbshipit-source-id: 1524fbea64461387fa640ae0123a21bf87a31e44