refactor container make_instance static method to function
Summary:
For thrift-python compatibility, we want to make the generated container types pure python classes. That means they can't have a cython static method to convert python container to cpp type.

We need to keep generating this method in some form to support use cases that need container to cpp converters. In auto-migrate mode, this will need to go in a separate converters module, but I'll leave that to a future diff.

Reviewed By: yoney

Differential Revision: D58383445

fbshipit-source-id: 82ecb9f97b9318295acf5fb4b47e69eba7262c8c