Clarify universal name docs
Summary:
Clarify that it is a user's responsibility to ensure uniqueness of universal names. We already do the same for package names which covers most cases but this covers the `thrift.Uri` case and is more explicit.

Replace "program" with "Thrift file" because the former is ambiguous in this context.

Improve the remaining sections and provide more meaningful examples similar to the ones in the IDL Reference.

Don't mention Conformance Any since it's a legacy API.

Swap the prefix and full hash columns in the table because the prefix one is cut off.

Fix a bunch of grammatical mistakes.

Reviewed By: avalonalex

Differential Revision: D58422633

fbshipit-source-id: aa10f69939099969f461890226fc1906cf6c64d8