Deshim PrimaryPtr in thrift
Summary:
The following rules were deshimmed:
```
//folly/experimental:primary_ptr -> //folly/concurrency/memory:primary_ptr
```

The following headers were deshimmed:
```
folly/experimental/PrimaryPtr.h -> folly/concurrency/memory/PrimaryPtr.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:deppw.thrift

Reviewed By: thedavekwon

Differential Revision: D58496410

fbshipit-source-id: aa10370d88822bc58d1dacb208da1e0dda162a45