Clear AssignPatch::data_ when AssignPatch::dynPatch_ is active when creating AssignPatch
Summary: AssignPatch currently has schema of `assign` and `clear`, but we only support `assign`. What this means is that in a rare case when `clear` operation  is specified AssignPatch created in Dynamic Patch, we might be having both AssignPatch::data_ and AssignPatch::dynPatch_ both active. In the normal case, this shouldn't matter, since logically AssignPatch::data_ and AssignPatch::dynPatch_ are union members, but when we `encode` with protocols that has kProtocolSupportsDynamicPatch false, we might bypass and use `AsignPatch::data_` which is problematic.

Reviewed By: Mizuchi

Differential Revision: D58539886

fbshipit-source-id: 0468e26b026ae52567679ea9a5e6110a98ebaf42