Include transitive dependencies in dep-map
Summary:
D58654187 is going to require that the Rust Thrift compiler has visibility into all transitive dependencies, not just direct dependencies of the Thrift library currently being compiled.

When a type needs to be mentioned for which the definition is not in any direct dependency, we can refer to it as `a::dependencies::b::dependencies::c` for some chain of dependencies a->b->c where `b` is a direct dependency of `a`, and `c` is a direct dependency of `b`.

This diff computes those dependency chains. Specifically, for each Thrift source file among the transitive dependencies of the current target, we compute the shortest chain of direct dependencies by which we can refer to it.

Reviewed By: zertosh

Differential Revision: D58656555

fbshipit-source-id: cfa7835b8ba63737af726f6c579d615325501c9c